+++ dio/lib/src/dio_exception.dart
@@ -173,6 +173,7 @@ class DioException implements Exception {
     required RequestOptions requestOptions,
     required String reason,
     Object? error,
+    StackTrace? stackTrace,
   }) =>
       DioException(
         type: DioExceptionType.connectionError,
@@ -181,6 +182,7 @@ class DioException implements Exception {
         requestOptions: requestOptions,
         response: null,
         error: error,
+        stackTrace: stackTrace,
       );
 
   /// The request info for the request that throws exception.